Implement addressable_gradient effect
This effect allows user defined gradients which the effect shifts through.
The gradient is defined with the gimp gradient format and a user defined length,
which should be much larger then the number of pixels in the strip. Setting flip
to true, will cause the gradient to change direction on the end.

fixes https://github.com/esphome/feature-requests/issues/417

Example:
```
light:
  - platform: neopixelbus
    type: GRBW
    variant: SK6812
    pin: RX
    num_leds: 20
    name: "Test neopixel"
    effects:
      - addressable_gradient:
          length: 50
          name: Fuschia 6
          use_white: true
          gradient: |
            GIMP Gradient
            Name: Fuschia 6
            4
            0.000000 0.125000 0.250000 0.811765 0.192157 0.870588 1.000000 0.784314 0.388235 0.937255 1.000000 0 0
            0.250000 0.375000 0.500000 0.784314 0.388235 0.937255 1.000000 0.560784 0.207843 0.964706 1.000000 0 0
            0.500000 0.625000 0.750000 0.560784 0.207843 0.964706 1.000000 0.823678 0.190171 0.822934 1.000000 0 0
            0.750000 0.875000 1.000000 0.823678 0.190171 0.822934 1.000000 0.560784 0.207843 0.964706 1.000000 0 0
```