Remove decode from str result, add type annotations
The str result of the call to `run_platformio_cli_run` had a decode call. Removed this as this is only required to convert bytes to str.
Put type return type annotation identifying the response of run_platformio_cli_run (and nested calls) as returning a Union of int or str.