wifi: retry connection if the connection is not valid
It was frequent that the ESP will report as "connected", but no
SSID/BSSID/IP were set. It will then disconnect imediatelly with
"Association Leave".
This ensures that a reconnect is done, which fixes the issue and
connects successfully.
This problem with the first connections may be related to this:
https://github.com/espressif/esp-idf/issues/2647