xiaomi_cgg1: add optional bindkey parameter.
Some versions of CGG1 appears to use the Xiaomi encrypted protocol,
rather than the unencrypted one.

Thankfully, they appear to be using exactly the same implementation as
everything else.

An alternative to this would be to match up the magic bytes to the CGD1
sensors, which they would be the same implementation of, but it would
be confusing having to choose a different platform for the sensors that
look exactly the same otherwise.