Refactor "cloud" connection helpers
Provide separate helpers for API and MQTT for targeted use and a
consolidating "remote" connection checker.

The consolidated helper will now return true if either type is connected
rather than both to cater for configs which make primary use of MQTT but
still have API configured for the sole purpose of the logger.