Prevent camera streaming from consuming the whole TCP window.
The existing code was adjusting the amount of camera data to be sent in
each cycle to be as much as would fit at the moment in the TCP send
window.

Unfortunately when streaming video that would fill the connection most
of the time, and important updates (sensor, light or switch state
changes) that happened to coincide with a camera data burst wouldn't
have space.

Given that the preexisting code wouldn't originally retry to send these
messages until the window got emptier, that meant that we usually lost
updates in Home Assistant.

This change adds a reserve of 64 bytes (typically enough to send a
simple update), and camera data will always leave that amount of buffer
available.