Better logging for API connection changes.
The following logging has been added to the api component:

**New client connections coming in**

In the original code, I saw that an attempt was done to log the client
information at connect time. However, this code was commented out, because
that piece of code ran in the lwIP thread, making the logging impossible.

I updated the code to log the incoming connection as soon as possible after
handling the connect. Implementation-wise I set a boolean at connect time,
which tells the api component to log the connect at its next `loop()`
iteration.

**Reasons for disconnecting an API client**

While debugging API client disconnect issues, I found that often no reason
was logged for disconnecting an API client. Therefore, I updated the api
component to have better logging. All disconnect routes now log a message,
making it a lot easier to determine where in the code and why the disconnect
was triggered.

**Remote port for connected clients**

I added the remote port to the client information that is shown in the log
messages.

Often, I have multiple connections open to the api server, from a single
source IP address. Without a port number, these connections cannot be
identified, and I cannot relate things that I see in the logs to an actual
process on the client side.

Knowing a port number also makes it a lot easier to setup a sniffer to
investigate the traffic that is being sent over the network for a specific
client connection.