Swap order of command and configuration file on command line
This gives a more natural command-line experience, in line with what
common utilities such as git and pip use.

It also requires a configuration file to be specified for commands that
require one, removes the need to specify one for commands that don't
need one, and doesn't allow to specify multiple configuration files for
commands that don't support that.