Drop 128x160 ESP-32 camera resolution (fixes esphome/issues#1884)
Since the addition of OV5640 support in the ESP-32 camera driver in
March 2020, it no longer provides the FRAMESIZE_QQVGA2 type. This was
imported in Arduino ESP32 core 1.0.5, so compilation was broken from
that version onwards.

Drop the support for this resolution, as this specific resolution does
not seem to be useful enough for it to be worthwhile to make it
conditionally available depending on the used Arduino version (which is
further complicated by the fact that Arduino doesn't provide a numeric
version define).