Add delays to work around I2C clock stretching limitations
The ESP32 I2C peripheral supports clock stretching of up to 13ms (when
APB clock is 80MHz). It seems that the SCD30 needs a longer clock
stretching particularly when setting configurations. However, it also
seems that this can be aleviated by adding delays in between
configurations. Let's just do exactly that.