Fix camera component on Arduino SDK 1.0.5+
It seems that in newer esp32-camera releases which are part of the
Arduino SDK 1.0.5/1.0.6 the frame size FRAMESIZE_QQVGA2 has beeen
removed. The reasons are unclear, and it seems that the frame size has
been removed/readded in the past already, so it might appear again one
day. Let's just support it if its defined in the SDK.