Fix brightness increasing during turn-off transition
Currently, the state and brightness are independently linearly
interpolated, and the product of them is written as the output. However,
when a light is turned off its brightness is set to 1 at that point.
Thus, the brightness of the turn-off transition for a light that
initially had brightness X follows the curve (1-t)*(X+(1-X)*t), which
for small values of X has a maximum larger than X (e.g. for initial
brightness 10%, the maximum brightness is 28%). This results in the
light actually getting brighter while turning it off - obviously not
desirable!

Instead, transition the brightness between its actual values, which
are zero when the light is off, and keep the state binary.

Fixes esphome/issues#1565.
Supersedes esphome/esphome#1643.