Fixed apparent race condition on ESP32 chips
I has a situation where when the cover first started moving the current position would often jump to a seemingly random value at the start. It seemed that the `this->last_recompute_time_ = now;` line was not getting called before `recompute_position_` was getting called. Once I moved this code up before the trigger this stopped been an issue. I haven't looked into the triggering code but perhaps this is due to the dual code of the ESP32?