Fix ethernet component hostname handling
Previously, the ethernet component would try to setup and connect before
`esp_eth_enable` is called, which can be called right after
`esp_eth_init` is called.  The rest of the logic in `start_connect_`,
for example setting up the hostname and dhcp information, should wait
until after the ETH_START event is received.  This would result in the
first connection attempt always failing (as `esp_eth_enable` would only
be called on the second run through `start_connect_`) and once
connected, would fail to set the hostname correctly.

This PR correctly runs `esp_eth_enable` and `esp_eth_init` in the
`setup` for the component.  This allows the ethernet component to
connect on the first attempt, saving 15 sec (the connection timeout) on
component startup.

This PR adds some additional state tracking to the ethernet component to
be able to know when the start event has been received and start the
connection once the eth_start event is received, and also track the
connection status through this state machine.  This makes the state more
explicit, rather than using `last_connected_` and `connected_` to track
state.

With this PR, because of the now correct setup, the hostname is
correctly set.  This could have been seen earlier by error checking the
return from `tcpip_adapter_set_hostname`, as this would fail on the
first call, before the esp_eth_enable.  Error checking is added to the
call to set the hostname, to catch further problems with this.

Fixes esphome/issues#2184