Correct ADS1115 handling of multiple sensors in continuous mode
When running in continuous mode, the ADS1115 config register will always
return that it is currently taking a measurement.  When the config
register needs to be set, either on the first time a given sensor is
read or if a different sensor with a different mux has been read on the
same device since the last update, then this will cause this function to
timeout.  When only using a single sensor on the ADS1115, this will only
cause the first sample to fail, and all other samples will be able to
immediately read the sensor value.  However, when there is more than one
sensor mux setup with the same update frequency, then every sensor
reading will fail, as the config will attempt to be updated on each
update.

To fix this, when running in continuous mode, simply assume that a
conversion has taken place after a sufficient amount of time has passed.
Testing this empirically, this is sufficient to always get the correct
voltage from the ADC. I checked this by configuring two sensors, one
measuring 0V and the other measuring at 3.3V, and both sensors always
reported the correct values within expected tolerances.  Ideally, we may
be able to use the ALERT/RDY pin on the ADS1115 to detect when the
conversion is complete, however this would take some additional
configuration and handling of the gpio pin polling or interrupt setup.

Fixes esphome/issues#892 (this was closed because it is stale)

There is some discussion in esphome/issues#892 about this problem, and
noting solutions similar to this, as well as the proposed fix using the
ready pin.  As noted in that issue, some of the issues mentioned there,
in particular regarding the CT clamp needing to poll the ADS1115 at
~kHz, will not be truly fixed by this, in particular because of the
inherent and necessary 2ms delay when changing mux channels (even with
an interrupt on the ready pin, this component will still need ~1.2ms to
register a new reading when running at 860 samples per second).
However, I feel like this is about as good as we can do given these
limitations and what this component is trying to accomplish. Some
documentation around the limitations here, as well as some information
in the CT clamp docs, may be helpful.