Remove superfluous polling on ADS1115
The `setup()` of the ADS1115 component would setup an update interval,
in addition to the one that would be setup by the polling component.
These calls would simply request a measurement but not update the value.
This additional call is removed by this change.  The ADS1115 continues
to function as expected using the polling setup by the PollingComponent.