Always tick mdns in ethernet component
Set the behavior of the ethernet component to mirror the wifi component
by always doing `network_tick_mdns` on each `loop()`.

The current implementation of `network_tick_mdns` is a noop for esp32,
meaning this function call will always do nothing as the ethernet
component is esp32-only.  Should any additional functionality be added
to this function for the esp32, this behavior is likely to be the most
correct, given the current configuration of the wifi component.

There is also an argument to be made to drop this call. If there is
never a case where esp32 specific functionality would be included in
`network_tick_mdns`, that may be a better change.