Fix Watchdog reset on ESP32
For ESP32 a Task Watchdog is used. However, yield() doesn't trigger the
task watchdog. The Task Watchdog is attached to the idle task, basically
making sure that the CPU isn't overloaded and goes into idle state at
least once every 5s. To force a Watchdog feed, we need to use a delay(1).