Reimplement TSL2591 using ESPHome I2C instead of the Adafruit library
This removes the dependency on the Adafruit TSL2591 library and the transitive
dependency on having an SPI configuration in ESPHome.

Along the way, I noticed that there was a rather large delay() in the
sensor readings in the Adafruit library. That's no problem for a simple
arduino-like framework, but it's a problem for ESPHome's update loop
for pollable components.

The root cause of the delay was allowing time for the ADCs to have
valid values after being enabled. And that came about because the device
was disabled after a reading and re-enabled for the next reading. That's
pretty good for power-saving, but still bad for ESPHome's update loop.

I made multiple integration changes because of that delay:

1. By default, I don't disable the device at all, so the ADC
readings don't need that delay. The ADCs operate continuously,
and you get the most recent completed integration cycle readings.

2. There is a configurable power saving mode. That gives the
disable/enable behavior that the Adafruit library has, but at the
cost of wonking out the ESPHome update loop. Docs recommend against
using the power saving mode, and it's off by default.

3. I exposed some additional C++ methods so that someone could
script their way through the enable-delay-read-disable process. Delays
in scripts are async, so they're not a problem for the ESPHome update
loop. (I have not tested doing this scripting.)