Tuya cover: use `force_set_datapoint_value()` for control datapoint
Force-sets the control datapoint, as the Tuya MCU doesn't send datapoint reports when the cover is set to a percentage or is controlled externally (manually, RF remote control, etc).  When the last reported value matches the command you're attempting to send, it was being skipped as the value was deemed to be unchanged.