Implement API over various transports
Existing TCP still works, obviously, but this adds
support for using a Stream (UART, currently) as a
transport.

This also adds an API client (only using a UART currently),
which will connect to an API Service and query the entities.
Given a suitable configuration file, the remote entities can
be adopted into the client ESPHome instance's entities as if
they were local. The primary reference for the remote entities
is the 'key' field, which is logged at the DEBUG or greater
level on the client instance. This is probably only available
over serial, due to the early stage at which the api_client
connects to the remote instance.

Configuration should look similar to:

  api_client:
   - id: uart2_api
     uart_id: uart2
     password: ''

  text_sensor:
   - platform: api_client
     key: 1234567
     name: Remote text sensor

where key is retrieved from the logs (or else manually calculated
using the rules and algorithms in esphome/core/component.h (class
Nameable, uint32_t get_object_id_hash()).

A future enhancement would be to store all api_client entities
defined in the configuration file in a temporary map, and move them
to their 'configured' state when the appropriate ListEntities*Response
is received. Then, when the ListEntitiesDoneResponse is received,
any remaining unmapped entities can be logged with a warning that
the remote entities appear not to exist, and therefore the local
entities will not function as expected.

Also updated the codegen to build a (currently non-existent)
AsyncAPIClientConnection, which takes a string-based IPv4 address,
and port as the target for the client connection.

Implementation of this code is still TODO. It will need to properly
manage disconnects, etc, exponential backoff, and such like to be a
good citizen.