Autodetect flash_size
With "Always upload using esptool (#2433)" ESPHome launches esptool.py
directly. It seems that this broke flashing of the ESP32-C3-DevKitM-1
which comes with 4MB flash. It seems that esptool.py is by default
configured to assume 2MB of flash.

Before, the Platform IO target "upload" has been used, which was passing
--flash_size detect as well. Replicate that behavior too.