Add reading secrets from .secrets.yaml
When using a ``secrets.yaml`` file, it always generates an ``invalid config`` message when using ``esphome <command> *yaml`` on the command line. This patch creates the possibility to use a ``.secrets.yaml`` file, so that it is hidden from the file system and isn't specified by the shell when using the ``*`` wildcard.