Add names of secrets files to constants
When ``esphome`` is run from the CLI to update a subset of the configuration files, or all of them, the ``secrets.yaml`` file - when specified - yields an error. This makes it hard to do, for instance, ``esphome upload *yaml``. This makes sure to skip the files ``secrets.yaml`` and ``secrets.yml`` when specified on the command line.