Extend `esp32_camera` with requester to improve performance
The `esp32_camera` has two consumers (requesters): web and api.
This improves a situation when two are concurrently used
to avoid situation when if the `web` requests stream or frame
all data is being feed via `api`. Instead it tracks
who requested stream or frame and allows receiver to make
a decision if to consume frame.

This also changes a way how stream is handled. Namely it does
introduce `start_stream` and `stop_stream` to control boundaries
when the stream is needed. Previously the `request_stream` would
pull frames for up-to 5 seconds after stream finished (in case of Web)
increasing CPU and power-usage. The 5 second stream re-request behavior
is retained for API which depends on periodically requesting streaming.

This changes how `idle_update_interval` is done. Previously the idle
request would be triggered an defined interval after the last frame
was requested. This changes to request the `Idle` unconditionally
every specific interval regardless of other requesters. This
makes us feed to `idle` frame every interval via API.

In general this improves performance and interoperability by:

- if Web is used, the frames are not feed into API, except periodic Idle frame
- if Web finishes streaming, and there are no more requesters the streaming of frames
  is immediately stopeed