Different run duration for non-timer wakeup
In some cases, it useful to set different run durations according to
the wakeup cause, which can be found in ESP32.

For example, a sleep duration of 10 minutes to conserve power. When
waking because of timer, only run for 10 seconds. However, if waking up
due to touch or gpio, stay awake for longer (be available for action,
wait for OTA, etc.)