Fix custom fan mode sorting
The UI presents fan modes in alphabetical order, which breaks the
ability to list numeric increments as the fan mode.

By prefixing each increment with leading spaces, the UI continues to
show the numbers as intended, but in the correct order. We still have
to compare the current custom fan mode using leading spaces.