SlowPWM: Fix switching output to a wrong state at the end of period
In rare occasions when `now` equals to `this->period_start_time_ + this->period_` (happens to me every couple of minutes in average on Sonoff 4CH with state=1.0 and period=10s) the output is set to false instead of always true (once for the whole period). Noticed that because my relay was clicking unexpectedly.