Eliminate delay(15) during setup() call, retry RESET commands 3x if failed
grep didn't find _any_ components using `set_retry`, but I think this is correct.
There's a little bit of the pyramid of doom and/or out-of-order code in `setup()`
due to scheduling future code and then current code showing up below it, but I feel okay
about this right now.