Add retries remaining parameter to the set_retry function
On the last attempt to successfully execute `f`, this will be 0, indicating there are no
more retries left.

Primary use case is so that `f` can customize behavior on the last execution, but there
may be some value in knowing how many are left, more than just a boolean indicating 0
or some.