Fix documentation: f is called immediately; fix code: first retry after `initial_wait_time`
My "rewrite" of the docs that claimed this doesn't execute `f` immediately was wrong, it
**is** being executed on the next scheduler loop, I was just trying to use an interval
that's less than the time for the initial setup tick to finish and didn't realize it.

This is still different from original docs in that the backoff_increase_factor default
was incorrectly documented, and I think specifying the increase happens via multiplication
is useful.

This commit *also* changes the code to match the originally documented backoff behavior.
As previously implemented, the 1st delay was (initial_wait_time * backoff_increase_factor),
contrary to the documentation. I think the as-documented behavior matches the method
signature the best, and I think this is a reasonably implemented bugfix.