Fix BLE280 setup when the sensor is marked as failed.
Some devices, like the t-higrow sensor from lilygo, power off all the sensors in order to save on batteries and then go on deep sleep.
When the esp comes back from deep sleep, the sensors are powered on again.
In the case of BLE280, it needs to be re-initialized, so the setup() method must be called.
This method should also clear the FAILED flag because, in case a read has been tried while the sensor was powered off, it would be permanently marked as failed.