The NEC IR protocol uses 8bit addresses and commands
The other bits are inverses for error checking.

This patch will gracefully accept 16bit addresses and commands for encoding but will emit a warning if you do. The decoder now only sets the proper 8bit address and command.

This patch will break existing implementations which expect to receive 16bit addresses and commands.

This patch will break existing implementations that rely on sending or receiving "made up" commands that fail the error checking.