Refactor BedJet climate into Hub component
This splits the BLE management out of the climate component, and into
a new hub component.

The climate component then registers itself as a child of the hub, in
order to receive status update notifications.

The hub exposes easy to use methods for sending simple commands to the
BedJet device, which can be used by other domain component child
devices, or from lambda actions/automations.