Show local MAC when scanning for the first time
When starting WiFi scan for the first time, the local MAC is not shown in the logs. When an AP has 'restrict adding new wireless devices' in its config (only connect to known MAC addresses), the ESP can't connect. You need to add the local MAC into the AP configuration, so you need this info.