Set humidity accuracy decimals to 1
The technical specification o fthe sensor has a resolution of 0.1%. I feel like the sensor should send the data as accurately as possible. Reduction of the resolution can then be done via `accuracy_decimals=0` in the configuration, if one is so inclined. As it stands, the higher resolution is intransparent to the user, as the documentation states: 

> "Manually set the number of decimals to use when reporting values. *This does not impact the actual value reported to Home Assistant, it just sets the number of decimals to use when displaying it.*"

The impression I got at first was that there would be no point in increasing the decimals, as there was no obvious way of knowing that there is already a default value in play that hides one digit of resolution.

I suggest either setting the default to 1 or stating in the documentation clearly, that one character of resolution is intentionally hidden.