Speed up BLE connections
We know enqueue the mtu and services search requests as
soon as the connection opens instead of waiting for the
mtu request to finish since they come back async. When
both have finished we tell the bluetooth proxy client
the state is now connected. This can shave a second
or two off the connect time.