Remove uuid lookups from BLE read/write/notify characteristics
Since the client is sending a specific handle they must
already have the service and char data. We want to be
able to release the memory used for the char/services
list in a future PR, but to do that we need to remove
some checks in places where we would end up reparsing
the list again.

See https://github.com/esphome/esphome/pull/4063/files#r1033010216