Combine the scanner failure check so we only stop the scanner once
If the scanner failed to start and the set scan param
failed we could try to stop the scanner twice to get
it back in a good state. We now check both at once
to ensure we only make one call to esp_ble_gap_stop_scanning
per loop.

Fix loop of
```
[10:18:04][E][esp32_ble_tracker:154]: Scan start failed: 1
[10:18:04][D][esp32_ble_tracker:289]: Starting scan...
[10:18:04][E][esp32_ble_tracker:148]: Scan set param failed: 1
[10:18:04][E][esp32_ble_tracker:154]: Scan start failed: 1
[10:18:04][D][esp32_ble_tracker:289]: Starting scan...
[10:18:04][E][esp32_ble_tracker:148]: Scan set param failed: 1
[10:18:04][E][esp32_ble_tracker:154]: Scan start failed: 1
```