Speed of bluetooth proxy connections with using esp-idf
We now enable CONFIG_BT_GATTC_CACHE_NVS_FLASH by default
unless cache_services is set to False when using esp-idf
since it allows us to avoid resolving the services on
every connection and store them on flash instead. This significantly
speeds up connecting to a device once we have already seen it.

As this will result in some additional wear on the flash
its only enabled when bluetooth proxies are enabled, and can
be disabled with the cache_services option