Remove unsafe strlen call on MQTT data
MQTT data is not necessarily a string, so this function may blindly
look for the first random null char in memory, which could result in
reading large blocks of unrelated memory and passing it to higher
layers as the message 'data'. The check for 0 assumes esp-mqtt was
going to give us invalid data_len, which appears unfounded. If data_len
is zero, we should not attempt to calculate this ourselves