Fix esp-idf MQTT backend holding dangling pointers
Simply copying the esp_mqtt_event_t in mqtt_event_handler was not
sufficient to be safe, since the data pointed to by topic, data and
error_handle was not guaranteed to be static by esp-mqtt. Instead,
the data in these fields must be copied before returning from the
event handler function. This was only generally noticeable on boot
when receiving retained messages, where it was likely that by the
time the event was handles, esp-mqtt had reused the underlying
buffer for the next event