Fix for I2C bus recovery w/scanning enabled
The long delay of 20*25ms (500ms) causes the WDT to reset
when the clock line is held low. This means the recovery
never completes. This change drastically reduces the
delay for bus recovery checks (HIGH) and resets the WDT
while attempting to do so. Total delay is roughly 1ms.
If the bus does not reset in a timely manner, it's
doubtful it will. Regardless, this avoids the WDT reset.