Setup output in Sigma Delta Output
The output pin, if configured, in the Sigma Delta Output was never initialized. This worked as long as the pin had been used as an output since power up since a software reset after an OTA didn't seem to reset the pin registers. Though I'm sure that last bit is probably expected behavior.