fix bug for xiaomi_miscale.cpp
After capture ble package and hack, there are some mistake in miscale code for XMTZC01HM. data[0] = 0x02 or 0x22 or 0xa2, it means kg unit, data[0] = 0x12 or 0x32 or 0xb2, it means jin unit and data[0] = 0x03 or 0x23 or 0xa3, it means lbs unit. For 0x02, 0x12, 0x03, it means in processing and the weight is unstable. For 0x22, 0x32, 0x23, it means processed and the weight is stable. And for 0xa2, 0xb2, 0xa3, it means the history data which is the earliest recent data as same as you measured in miscale. So in code, I just to choose the stable data for record. And the second, for jin unit, I don't know why to * 0.6f. In my device, it is not needed.